+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v48.1.0 (2023-03-27)
+
+### Added
+
+- (bugsnag-expo-cli) Updated bugsnag-expo-cli to support Typescript [#98](https://github.com/bugsnag/bugsnag-expo/pull/98)
+- Read API key and app version from `Constants.expoConfig` [#119](https://github.com/bugsnag/bugsnag-expo/pull/119)
+
+### Fixed
+
+- (plugin-expo-eas-sourcemaps) Reinstate API key in Android manifest [#117](https://github.com/bugsnag/bugsnag-expo/pull/117)
+- (plugin-expo-eas-sourcemaps) Support dynamic configuration files in EAS Build lifecycle hook [#117](https://github.com/bugsnag/bugsnag-expo/pull/117)
+
 ## v48.0.0 (2023-03-07)
 
 This release adds support for expo 48